pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "tower"
version = "0.3.10"
description = "Tower CLI and runtime environment for Tower."
authors = [{ name = "Tower Computing Inc.", email = "brad@tower.dev" }]
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
keywords = [
  "automation",
  "flake8",
  "pycodestyle",
  "pyflakes",
  "pylint",
  "clippy",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Rust",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
  "attrs>=24.2.0",
  "httpx>=0.23.3",
  "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
ai = ["huggingface-hub>=0.30.2", "ollama>=0.4.7"]
iceberg = ["polars>=1.27.1", "pyarrow>=19.0.1", "pyiceberg>=0.9.0"]
all = ["tower[ai,iceberg]"]
dev = [
	"openapi-python-client>=0.12.1",
  "pytest>=8.3.5",
  "pytest-httpx>=0.35.0",
]

[tool.maturin]
bindings = "bin"
manifest-path = "crates/tower/Cargo.toml"
module-name = "tower"
python-source = "src"
strip = true
exclude = []
include = ["rust-toolchain.toml"]

[tool.uv.sources]
tower = { workspace = true }
